Added the attendance calculation logic. Clean up of old code. New exceptions added. Started working on services.

+++ src/main/java/com/ciesielski/schoolapi/domain/model/School.java
@@ -3,10 +3,12 @@ package com.ciesielski.schoolapi.domain.model;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class School {
     private Long id;
     private String name;
-    private double hour_price;
+    private BigDecimal hour_price; // todo: precision 2
 }
